requirements.txt
Babel==2.6.0
certifi==2018.8.13
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1
django-allauth==0.36.0
django-filter==2.0.0
django-phonenumber-field==2.0.1
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-gis==0.13
djangorestframework-jwt==1.11.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
openapi-codec==1.3.2
phonenumbers==8.9.10
psycopg2==2.7.5
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
